requirements.txt
# Core LLM and Agent Framework
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.1.0
langgraph>=0.1.0

# Judgment Labs SDK
judgeval>=0.0.50

# Search and APIs
tavily-python>=0.3.0

# Environment and Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0